organizo los ficheros y subo la funcion custom env
