+++ 0_cosas/apuntes.md
