+++ 0_cosas/dudas.md
@@ -1,0 +1,3 @@
+# 1
+para que es el ultimo int de esta funcion
+/Users/msintas-/Documents/Cursus42/MiniShell/Minishell-new/dbl_utils/ft_dbl_printf.c
