+++ 0_cosas/ejemplos-comandos.md
