+++ 0_cosas/fallos-pendientes.md
