+++ 1_pruebas/new_shell_split.c
