+++ builtins/custom_unset.c
@@ -1,1 +1,19 @@
 #include "../minishell.h"
+/* empezada por acaraban */
+
+/*int unset(t_content *cont, int i)
+{
+	int j;
+	int par;
+
+	j = 0;
+	par = 0;
+	if (!strcmp(cont[i].cmd, "unset"))
+	{
+		while (env[j] && !par)
+			par = str_cmp(cont[i].full_comand[1], 0, cont->global->env[j++], '=');
+		if (par)
+			
+	}
+	return (0);
+}*/
