--- builtins/unset.c
@@ -1,16 +1,0 @@
-int unset(t_content *cont, int i)
-{
-	int j;
-	int par;
-
-	j = 0;
-	par = 0;
-	if (!strcmp(cont[i].cmd, "unset"))
-	{
-		while (env[j] && !par)
-			par = str_cmp(cont[i].full_comand[1], 0, cont->global->env[j++], '=');
-		if (par)
-			
-	}
-	return (0);
-}
