+++ dbl_utils/ft_dbl_printf.c
@@ -1,5 +1,13 @@
 #include "../minishell.h"
 
+/*
+	This function is to print a double pointer.
+	txt: 	text before the array item
+	ftxt:	text after
+	sal:	0 -> ?
+			1 -> ? 
+*/
+
 void	ft_dbl_printf(char *txt, char **arr, char *ftxt, int sal)
 {
 	int i;
