--- heredoc.c
@@ -1,8 +1,0 @@
-#include "minishell.h"
-
-char *ft_heredoc(t_content *cont)
-{
-	char *val;
-
-	while ()
-}
